\name{so_Bayesian}
\alias{so_Bayesian}
\title{so_Bayesian reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PopulationEstimates/Bayesian element of a PharmML-SO data structure
}
\section{Methods}{
so_Bayesian$new() - Create a new empty so_Bayesian object\cr
}
\section{Fields}{
$PosteriorMean - A data.frame\cr
$PosteriorMedian - A data.frame\cr
$PosteriorMode - A data.frame\cr
}
\keyword{so_Bayesian}
