\name{so_Bayesian_PPE}
\alias{so_Bayesian_PPE}
\title{so_Bayesian_PPE reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionPopulationEstimates/Bayesian element of a PharmML-SO data structure
}
\section{Methods}{
so_Bayesian_PPE$new() - Create a new empty so_Bayesian_PPE object\cr
}
\section{Fields}{
$StandardDeviationPosterior - A data.frame\cr
$PercentilesCI - A data.frame\cr
}
\keyword{so_Bayesian_PPE}
