\name{so_Bootstrap}
\alias{so_Bootstrap}
\title{so_Bootstrap reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PopulationEstimates/Bootstrap element of a PharmML-SO data structure
}
\section{Methods}{
so_Bootstrap$new() - Create a new empty so_Bootstrap object
}
\section{Fields}{
$Mean - A data.frame\cr
$Median - A data.frame\cr
}
\keyword{so_Bootstrap}
