\name{so_ConditionNumber}
\alias{so_ConditionNumber}
\title{so_ConditionNumber reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionPopulationEstimates/ConditionNumber element of a PharmML-SO data structure
}
\section{Methods}{
so_ConditionNumber$new() - Create a new empty so_ConditionNumber object
}
\section{Fields}{
$Real - A numeric\cr
}
\keyword{so_ConditionNumber}
