\name{so_Content}
\alias{so_Content}
\title{so_Content reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/Message/Content element of a PharmML-SO data structure
}
\section{Methods}{
so_Content$new() - Create a new empty so_Content object
}
\section{Fields}{
$String - A character string\cr
}
\keyword{so_Content}
