\name{so_DataFile}
\alias{so_DataFile}
\title{so_DataFile reference class}
\description{
	Reference Class for the SO/SOBlock/RawResults/DataFile element of a PharmML-SO data structure
}
\section{Methods}{
so_DataFile$new() - Create a new empty so_DataFile object
}
\section{Fields}{
$Description - A character string\cr
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_DataFile}
