\name{so_Design}
\alias{so_Design}
\title{so_Design reference class}
\description{
	Reference Class for the SO/SOBlock/OptimalDesign/OptimalDesignBlock/Design element of a PharmML-SO data structure
}
\section{Methods}{
so_Design$new() - Create a new empty so_Design object
}
\section{Fields}{
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_Design}
