\name{so_DiagnosticIndividualParams}
\alias{so_DiagnosticIndividualParams}
\title{so_DiagnosticIndividualParams reference class}
\description{
	Reference Class for the SO/SOBlock/ModelDiagnostic/DiagnosticIndividualParams element of a PharmML-SO data structure
}
\section{Methods}{
so_DiagnosticIndividualParams$new() - Create a new empty so_DiagnosticIndividualParams object\cr
}
\section{Fields}{
$RandomEffects - A data.frame\cr
$IndivParamsCovariates - A data.frame\cr
$DistributionIndivParams - A data.frame\cr
}
\keyword{so_DiagnosticIndividualParams}
