\name{so_DiagnosticPlotsStructuralModel}
\alias{so_DiagnosticPlotsStructuralModel}
\title{so_DiagnosticPlotsStructuralModel reference class}
\description{
	Reference Class for the SO/SOBlock/ModelDiagnostic/DiagnosticPlotsStructuralModel element of a PharmML-SO data structure
}
\section{Methods}{
so_DiagnosticPlotsStructuralModel$new() - Create a new empty so_DiagnosticPlotsStructuralModel object
}
\section{Fields}{
$IndivFits - A \link{so_IndivFits} object\cr
$IndivPredictionVsObserv - A data.frame\cr
$VPC - A data.frame\cr
}
\keyword{so_DiagnosticPlotsStructuralModel}
