\name{so_DiagnosticStructuralModel}
\alias{so_DiagnosticStructuralModel}
\title{so_DiagnosticStructuralModel reference class}
\description{
	Reference Class for the SO/SOBlock/ModelDiagnostic/DiagnosticStructuralModel element of a PharmML-SO data structure
}
\section{Methods}{
so_DiagnosticStructuralModel$new() - Create a new empty so_DiagnosticStructuralModel object\cr
}
\section{Fields}{
$IndivObservationPrediction - A data.frame\cr
$VPC - A data.frame\cr
}
\keyword{so_DiagnosticStructuralModel}
