\name{dv_column}
\alias{dv_column}
\title{Get index of the DV column in a data.frame created from an SO object}
\description{
	Function to get the index of the dependent variable from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{dv_column}
