\name{dv_column_name}
\alias{dv_column_name}
\title{Get the name of the DV column in a data.frame created from an SO object}
\description{
	Function to get the name of the dependent variable from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{dv_column_name}
