\name{so_Estimates}
\alias{so_Estimates}
\title{so_Estimates reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/IndividualEstimates/Estimates element of a PharmML-SO data structure
}
\section{Methods}{
so_Estimates$new() - Create a new empty so_Estimates object\cr
}
\section{Fields}{
$Mean - A data.frame\cr
$Median - A data.frame\cr
$Mode - A data.frame\cr
$Samples - A data.frame\cr
}
\keyword{so_Estimates}
