\name{so_Estimation}
\alias{so_Estimation}
\title{so_Estimation reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation element of a PharmML-SO data structure
}
\section{Methods}{
so_Estimation$new() - Create a new empty so_Estimation object\cr
}
\section{Fields}{
$PopulationEstimates - A \link{so_PopulationEstimates} object\cr
$PrecisionPopulationEstimates - A \link{so_PrecisionPopulationEstimates} object\cr
$IndividualEstimates - A \link{so_IndividualEstimates} object\cr
$PrecisionIndividualEstimates - A \link{so_PrecisionIndividualEstimates} object\cr
$Residuals - A \link{so_Residuals} object\cr
$Predictions - A data.frame\cr
$OFMeasures - A \link{so_OFMeasures} object\cr
$TargetToolMessages - A \link{so_TargetToolMessages} object\cr
}
\keyword{so_Estimation}
