\name{so_File}
\alias{so_File}
\title{so_File reference class}
\description{
	Reference Class for the SO/SOBlock/ToolSettings/File element of a PharmML-SO data structure
}
\section{Methods}{
so_File$new() - Create a new empty so_File object
}
\section{Fields}{
$Description - A character string\cr
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_File}
