\name{so_GraphicsFile}
\alias{so_GraphicsFile}
\title{so_GraphicsFile reference class}
\description{
	Reference Class for the SO/SOBlock/RawResults/GraphicsFile element of a PharmML-SO data structure
}
\section{Methods}{
so_GraphicsFile$new() - Create a new empty so_GraphicsFile object
}
\section{Fields}{
$Description - A character string\cr
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_GraphicsFile}
