\name{id_column}
\alias{id_column}
\title{Get index of the ID column in a data.frame created from an SO object}
\description{
	Function to get the index of the ID column from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{id_column}
