\name{id_column_name}
\alias{id_column_name}
\title{Get the name of the ID column in a data.frame created from an SO object}
\description{
	Function to get the name of the ID column from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{id_column_name}
