\name{idv_column}
\alias{idv_column}
\title{Get index of the independent variable column in a data.frame created from an SO object}
\description{
	Function to get the index of the independent variable from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{idv_column}
