\name{idv_column_name}
\alias{idv_column_name}
\title{Get the name of the independent variable column in a data.frame created from an SO object}
\description{
	Function to get the name of the independent variable from a data.frame created from an SO object
}
\arguments{
    \item{table}{A data.frame extracted from an SO object}
}
\keyword{idv_column_name}
