\name{so_IndependentVariable}
\alias{so_IndependentVariable}
\title{so_IndependentVariable reference class}
\description{
	Reference Class for the PharmML/IndependentVariable element of a PharmML-SO data structure
}
\section{Methods}{
so_IndependentVariable$new() - Create a new empty so_IndependentVariable object
}
\section{Fields}{
$Description - A character string\cr
$id - A character string attribute\cr
$symbId - A character string attribute\cr
}
\keyword{so_IndependentVariable}
