\name{so_IndivFits}
\alias{so_IndivFits}
\title{so_IndivFits reference class}
\description{
	Reference Class for the SO/SOBlock/ModelDiagnostic/DiagnosticPlotsStructuralModel/IndivFits element of a PharmML-SO data structure
}
\section{Methods}{
so_IndivFits$new() - Create a new empty so_IndivFits object
}
\section{Fields}{
$ObservationTable - A data.frame\cr
$PredictionTable - A data.frame\cr
}
\keyword{so_IndivFits}
