\name{so_IndividualEstimates}
\alias{so_IndividualEstimates}
\title{so_IndividualEstimates reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/IndividualEstimates element of a PharmML-SO data structure
}
\section{Methods}{
so_IndividualEstimates$new() - Create a new empty so_IndividualEstimates object\cr
}
\section{Fields}{
$Estimates - A \link{so_Estimates} object\cr
$RandomEffects - A \link{so_RandomEffects_IE} object\cr
$EtaShrinkage - A data.frame\cr
}
\keyword{so_IndividualEstimates}
