\name{so_InformationCriteria}
\alias{so_InformationCriteria}
\title{so_InformationCriteria reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/OFMeasures/InformationCriteria element of a PharmML-SO data structure
}
\section{Methods}{
so_InformationCriteria$new() - Create a new empty so_InformationCriteria object\cr
}
\section{Fields}{
$AIC - A numeric\cr
$BIC - A numeric\cr
$DIC - A numeric\cr
}
\keyword{so_InformationCriteria}
