\name{so_Level}
\alias{so_Level}
\title{so_Level reference class}
\description{
	Reference Class for the PharmML/ModelDefinition/VariabilityModel/Level element of a PharmML-SO data structure
}
\section{Methods}{
so_Level$new() - Create a new empty so_Level object
}
\section{Fields}{
$Description - A character string\cr
$Name - A character string\cr
$ParentLevel - A \link{so_ParentLevel} object\cr
$id - A character string attribute\cr
$referenceLevel - A character string attribute\cr
$symbId - A character string attribute\cr
}
\keyword{so_Level}
