\name{so_Likelihood}
\alias{so_Likelihood}
\title{so_Likelihood reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/Likelihood element of a PharmML-SO data structure
}
\section{Methods}{
so_Likelihood$new() - Create a new empty so_Likelihood object
}
\section{Fields}{
$LogLikelihood - A numeric\cr
$Deviance - A numeric\cr
$IndividualContribtoLL - A data.frame\cr
}
\keyword{so_Likelihood}
