\name{so_Message}
\alias{so_Message}
\title{so_Message reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/Message element of a PharmML-SO data structure
}
\section{Methods}{
so_Message$new() - Create a new empty so_Message object\cr
}
\section{Fields}{
$Toolname - A character string\cr
$Name - A character string\cr
$Content - A character string\cr
$Severity - An integer\cr
$type - A character string attribute\cr
}
\keyword{so_Message}
