\name{so_MissingData}
\alias{so_MissingData}
\title{so_MissingData reference class}
\description{
	Reference Class for the MissingData element of a PharmML-SO data structure
}
\section{Methods}{
so_MissingData$new() - Create a new empty so_MissingData object\cr
}
\section{Fields}{
$dataCode - A character string attribute\cr
$missingDataType - A character string attribute\cr
}
\keyword{so_MissingData}
