\name{so_MLE}
\alias{so_MLE}
\title{so_MLE reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionPopulationEstimates/MLE element of a PharmML-SO data structure
}
\section{Methods}{
so_MLE$new() - Create a new empty so_MLE object\cr
}
\section{Fields}{
$FIM - A matrix\cr
$CovarianceMatrix - A matrix\cr
$CorrelationMatrix - A matrix\cr
$StandardError - A data.frame\cr
$RelativeStandardError - A data.frame\cr
$AsymptoticCI - A data.frame\cr
$ConditionNumber - A numeric\cr
}
\keyword{so_MLE}
