\name{so_ModelDefinition}
\alias{so_ModelDefinition}
\title{so_ModelDefinition reference class}
\description{
	Reference Class for the PharmML/ModelDefinition element of a PharmML-SO data structure
}
\section{Methods}{
so_ModelDefinition$new() - Create a new empty so_ModelDefinition object
}
\section{Fields}{
$VariabilityModel - A \link{so_VariabilityModel} object\cr
}
\keyword{so_ModelDefinition}
