\name{so_ModelDiagnostic}
\alias{so_ModelDiagnostic}
\title{so_ModelDiagnostic reference class}
\description{
	Reference Class for the SO/SOBlock/ModelDiagnostic element of a PharmML-SO data structure
}
\section{Methods}{
so_ModelDiagnostic$new() - Create a new empty so_ModelDiagnostic object\cr
}
\section{Fields}{
$DiagnosticStructuralModel - A \link{so_DiagnosticStructuralModel} object\cr
$DiagnosticIndividualParams - A \link{so_DiagnosticIndividualParams} object\cr
}
\keyword{so_ModelDiagnostic}
