\name{so_Name}
\alias{so_Name}
\title{so_Name reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/Message/Name element of a PharmML-SO data structure
}
\section{Methods}{
so_Name$new() - Create a new empty so_Name object
}
\section{Fields}{
$String - A character string\cr
}
\keyword{so_Name}
