\name{so_NumberChains}
\alias{so_NumberChains}
\title{so_NumberChains reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/NumberChains element of a PharmML-SO data structure
}
\section{Methods}{
so_NumberChains$new() - Create a new empty so_NumberChains object
}
\section{Fields}{
$Description - A character string\cr
$Real - A numeric\cr
}
\keyword{so_NumberChains}
