\name{so_NumberIterations}
\alias{so_NumberIterations}
\title{so_NumberIterations reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/NumberIterations element of a PharmML-SO data structure
}
\section{Methods}{
so_NumberIterations$new() - Create a new empty so_NumberIterations object
}
\section{Fields}{
$Real - A numeric\cr
}
\keyword{so_NumberIterations}
