\name{so_OFMeasures}
\alias{so_OFMeasures}
\title{so_OFMeasures reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/OFMeasures element of a PharmML-SO data structure
}
\section{Methods}{
so_OFMeasures$new() - Create a new empty so_OFMeasures object\cr
}
\section{Fields}{
$Likelihood - A numeric\cr
$LogLikelihood - A numeric\cr
$Deviance - A numeric\cr
$ToolObjFunction - A numeric\cr
$IndividualContribToLL - A data.frame\cr
$InformationCriteria - A \link{so_InformationCriteria} object\cr
}
\keyword{so_OFMeasures}
