\name{so_OptimalDesign}
\alias{so_OptimalDesign}
\title{so_OptimalDesign reference class}
\description{
	Reference Class for the SO/SOBlock/OptimalDesign element of a PharmML-SO data structure
}
\section{Methods}{
so_OptimalDesign$new() - Create a new empty so_OptimalDesign object\cr
so_OptimalDesign$add_OptimalDesignBlock(object) - Add a OptimalDesignBlock\cr
so_OptimalDesign$remove_OptimalDesignBlock(object, i) - Remove the OptimalDesignBlock having index i\cr
}
\section{Fields}{
$OptimalDesignBlock - A list of \link{so_OptimalDesignBlock} objects\cr
$type - A character string attribute\cr
}
\keyword{so_OptimalDesign}
