\name{so_OptimalDesignBlock}
\alias{so_OptimalDesignBlock}
\title{so_OptimalDesignBlock reference class}
\description{
	Reference Class for the SO/SOBlock/OptimalDesign/OptimalDesignBlock element of a PharmML-SO data structure
}
\section{Methods}{
so_OptimalDesignBlock$new() - Create a new empty so_OptimalDesignBlock object\cr
}
\section{Fields}{
$FIM - A matrix\cr
$CovarianceMatrix - A matrix\cr
$ParameterPrecision - A data.frame\cr
$Criteria - A data.frame\cr
$Tests - A data.frame\cr
$SimulatedData - A \link{so_ExternalFile} object\cr
$Design - A \link{so_ExternalFile} object\cr
$blockNumber - An integer attribute\cr
}
\keyword{so_OptimalDesignBlock}
