\name{so_OtherMethod}
\alias{so_OtherMethod}
\title{so_OtherMethod reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PopulationEstimates/OtherMethod element of a PharmML-SO data structure
}
\section{Methods}{
so_OtherMethod$new() - Create a new empty so_OtherMethod object\cr
}
\section{Fields}{
$Mean - A data.frame\cr
$Median - A data.frame\cr
$method - A character string attribute\cr
}
\keyword{so_OtherMethod}
