\name{so_OtherMethod_PPE}
\alias{so_OtherMethod_PPE}
\title{so_OtherMethod_PPE reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionPopulationEstimates/OtherMethod element of a PharmML-SO data structure
}
\section{Methods}{
so_OtherMethod_PPE$new() - Create a new empty so_OtherMethod_PPE object\cr
}
\section{Fields}{
$CovarianceMatrix - A matrix\cr
$CorrelationMatrix - A matrix\cr
$StandardDeviation - A data.frame\cr
$StandardError - A data.frame\cr
$AsymptoticCI - A data.frame\cr
$PercentilesCI - A data.frame\cr
}
\keyword{so_OtherMethod_PPE}
