\name{so_OutputFilePath}
\alias{so_OutputFilePath}
\title{so_OutputFilePath reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/OutputFilePath element of a PharmML-SO data structure
}
\section{Methods}{
so_OutputFilePath$new() - Create a new empty so_OutputFilePath object
}
\section{Fields}{
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_OutputFilePath}
