\name{so_ParentLevel}
\alias{so_ParentLevel}
\title{so_ParentLevel reference class}
\description{
	Reference Class for the PharmML/ModelDefinition/VariabilityModel/Level/ParentLevel element of a PharmML-SO data structure
}
\section{Methods}{
so_ParentLevel$new() - Create a new empty so_ParentLevel object
}
\section{Fields}{
$Description - A character string\cr
$SymbRef - A \link{so_SymbRef} object\cr
$id - A character string attribute\cr
}
\keyword{so_ParentLevel}
