\name{so_PharmML}
\alias{so_PharmML}
\title{so_PharmML reference class}
\description{
	Reference Class for the PharmML element of a PharmML-SO data structure
}
\section{Methods}{
so_PharmML$new() - Create a new empty so_PharmML object
}
\section{Fields}{
$Name - A character string\cr
$IndependentVariable - A \link{so_IndependentVariable} object\cr
$ModelDefinition - A \link{so_ModelDefinition} object\cr
}
\keyword{so_PharmML}
