\name{so_PharmMLRef}
\alias{so_PharmMLRef}
\title{so_PharmMLRef reference class}
\description{
	Reference Class for the SO/PharmMLRef element of a PharmML-SO data structure
}
\section{Methods}{
so_PharmMLRef$new() - Create a new empty so_PharmMLRef object\cr
}
\section{Fields}{
$Description - A character string\cr
$name - A character string attribute\cr
$id - A character string attribute\cr
}
\keyword{so_PharmMLRef}
