\name{so_PopulationEstimates}
\alias{so_PopulationEstimates}
\title{so_PopulationEstimates reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PopulationEstimates element of a PharmML-SO data structure
}
\section{Methods}{
so_PopulationEstimates$new() - Create a new empty so_PopulationEstimates object\cr
}
\section{Fields}{
$MLE - A data.frame\cr
$Bayesian - A \link{so_Bayesian} object\cr
$OtherMethod - A \link{so_OtherMethod} object\cr
}
\keyword{so_PopulationEstimates}
