\name{so_PrecisionIndividualEstimates}
\alias{so_PrecisionIndividualEstimates}
\title{so_PrecisionIndividualEstimates reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionIndividualEstimates element of a PharmML-SO data structure
}
\section{Methods}{
so_PrecisionIndividualEstimates$new() - Create a new empty so_PrecisionIndividualEstimates object\cr
}
\section{Fields}{
$StandardDeviation - A data.frame\cr
$PercentilesCI - A data.frame\cr
}
\keyword{so_PrecisionIndividualEstimates}
