\name{so_PrecisionPopulationEstimates}
\alias{so_PrecisionPopulationEstimates}
\title{so_PrecisionPopulationEstimates reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/PrecisionPopulationEstimates element of a PharmML-SO data structure
}
\section{Methods}{
so_PrecisionPopulationEstimates$new() - Create a new empty so_PrecisionPopulationEstimates object\cr
}
\section{Fields}{
$MLE - A \link{so_MLE} object\cr
$Bayesian - A \link{so_Bayesian_PPE} object\cr
$OtherMethod - A \link{so_OtherMethod_PPE} object\cr
}
\keyword{so_PrecisionPopulationEstimates}
