\name{so_RandomEffects_IE}
\alias{so_RandomEffects_IE}
\title{so_RandomEffects_IE reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/IndividualEstimates/RandomEffects element of a PharmML-SO data structure
}
\section{Methods}{
so_RandomEffects_IE$new() - Create a new empty so_RandomEffects_IE object\cr
}
\section{Fields}{
$EffectMean - A data.frame\cr
$EffectMedian - A data.frame\cr
$EffectMode - A data.frame\cr
$Samples - A data.frame\cr
}
\keyword{so_RandomEffects_IE}
