\name{so_Residuals}
\alias{so_Residuals}
\title{so_Residuals reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/Residuals element of a PharmML-SO data structure
}
\section{Methods}{
so_Residuals$new() - Create a new empty so_Residuals object\cr
}
\section{Fields}{
$ResidualTable - A data.frame\cr
$EpsShrinkage - A data.frame\cr
}
\keyword{so_Residuals}
