\name{reverse-package}
\alias{reverse-package}
\docType{package}
\title{
Reverse an integer vector
}
\description{
A function to reverse a vector.
}
\details{
\tabular{ll}{
Package: \tab reverse\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-05\cr
License: \tab BSD-3\cr
}
}
\author{
Bjorn Winckler

Maintainer: Bjorn Winckler <bjorn.winckler@gmail.com>
}
\keyword{ package }
\examples{
}
