\name{reverse}
\alias{reverse}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reverse an integer vector}
\description{
    A function that reverses an integer vector.
}
\usage{
reverse(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    An integer vector.
}
}
\value{
    The input vector reversed.
}
\author{
    Bjorn Winckler
}

\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{reverse}
