\name{so_RunTime}
\alias{so_RunTime}
\title{so_RunTime reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/RunTime element of a PharmML-SO data structure
}
\section{Methods}{
so_RunTime$new() - Create a new empty so_RunTime object
}
\section{Fields}{
$Description - A character string\cr
$Real - A numeric\cr
}
\keyword{so_RunTime}
