\name{so_Severity}
\alias{so_Severity}
\title{so_Severity reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/Message/Severity element of a PharmML-SO data structure
}
\section{Methods}{
so_Severity$new() - Create a new empty so_Severity object
}
\section{Fields}{
$Int - An integer\cr
}
\keyword{so_Severity}
