\name{so_SimulatedData}
\alias{so_SimulatedData}
\title{so_SimulatedData reference class}
\description{
	Reference Class for the SO/SOBlock/OptimalDesign/OptimalDesignBlock/SimulatedData element of a PharmML-SO data structure
}
\section{Methods}{
so_SimulatedData$new() - Create a new empty so_SimulatedData object
}
\section{Fields}{
$path - A character string\cr
$oid - A character string attribute\cr
}
\keyword{so_SimulatedData}
