\name{so_Simulation}
\alias{so_Simulation}
\title{so_Simulation reference class}
\description{
	Reference Class for the SO/SOBlock/Simulation element of a PharmML-SO data structure
}
\section{Methods}{
so_Simulation$new() - Create a new empty so_Simulation object\cr
so_Simulation$add_SimulationBlock(object) - Add a SimulationBlock\cr
so_Simulation$remove_SimulationBlock(object, i) - Remove the SimulationBlock having index i\cr
}
\section{Fields}{
$SimulationBlock - A list of \link{so_SimulationBlock} objects\cr
}
\keyword{so_Simulation}
