\name{so_SimulationSubType}
\alias{so_SimulationSubType}
\title{so_SimulationSubType reference class}
\description{
	Reference Class for the SimulationSubType element of a PharmML-SO data structure
}
\section{Methods}{
so_SimulationSubType$new() - Create a new empty so_SimulationSubType object\cr
}
\section{Fields}{
$name - A character string attribute\cr
$extFileNo - An integer attribute\cr
}
\keyword{so_SimulationSubType}
