\name{so_SOBlock}
\alias{so_SOBlock}
\title{so_SOBlock reference class}
\description{
	Reference Class for the SO/SOBlock element of a PharmML-SO data structure
}
\section{Methods}{
so_SOBlock$new() - Create a new empty so_SOBlock object\cr
}
\section{Fields}{
$ToolSettings - A \link{so_ToolSettings} object\cr
$RawResults - A \link{so_RawResults} object\cr
$TaskInformation - A \link{so_TaskInformation} object\cr
$Estimation - A \link{so_Estimation} object\cr
$Simulation - A \link{so_Simulation} object\cr
$ModelDiagnostic - A \link{so_ModelDiagnostic} object\cr
$OptimalDesign - A \link{so_OptimalDesign} object\cr
$blkId - A character string attribute\cr
}
\keyword{so_SOBlock}
