\name{soc-package}
\alias{soc-package}
\docType{package}
\title{
A package to handle PharmML standardized output (SO) xml files
}
\description{
A package to handle PharmML standardized output (SO) xml files.
SO files can be created, read, manipulated and written through a
data binding from the XML structure to a tree structure of R objects.
}
\details{
\tabular{ll}{
Package: \tab soc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-11-24\cr
License: \tab LGPL-3\cr
}
}
\author{
Rikard Nordgren

Maintainer: Rikard Nordgren <rikard.nordgren@farmbio.uu.se>
}
\keyword{package}
\examples{
}
