\name{so_Bootstrap}
\alias{so_Bootstrap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{so_Bootstrap title}
\description{
    Class for SO/SOBlock/Estimation/PopulationParameters/Bootstrap
}
\usage{
so_Bootstrap
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    An integer vector.
}
}
\value{
    The input vector reversed.
}
\author{
    Rikard Nordgren
}

\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{so_Bootstrap}
