\name{so_SymbRef}
\alias{so_SymbRef}
\title{so_SymbRef reference class}
\description{
	Reference Class for the SymbRef element of a PharmML-SO data structure
}
\section{Methods}{
so_SymbRef$new() - Create a new empty so_SymbRef object
}
\section{Fields}{
$Description - A character string\cr
$id - A character string attribute\cr
$blkIdRef - A character string attribute\cr
$symbIdRef - A character string attribute\cr
}
\keyword{so_SymbRef}
