\name{so_TargetToolMessages}
\alias{so_TargetToolMessages}
\title{so_TargetToolMessages reference class}
\description{
	Reference Class for the SO/SOBlock/Estimation/TargetToolMessages element of a PharmML-SO data structure
}
\section{Methods}{
so_TargetToolMessages$new() - Create a new empty so_TargetToolMessages object\cr
}
\section{Fields}{
$Termination - A character string\cr
$Warnings - A character string\cr
$Errors - A character string\cr
$ElapsedTime - A numeric\cr
$OutputFilePath - A \link{so_ExternalFile} object\cr
$ChainsNumber - A numeric\cr
$IterationNumber - A numeric\cr
}
\keyword{so_TargetToolMessages}
