\name{so_Toolname}
\alias{so_Toolname}
\title{so_Toolname reference class}
\description{
	Reference Class for the SO/SOBlock/TaskInformation/Message/Toolname element of a PharmML-SO data structure
}
\section{Methods}{
so_Toolname$new() - Create a new empty so_Toolname object
}
\section{Fields}{
$String - A character string\cr
}
\keyword{so_Toolname}
