\name{so_ToolSettings}
\alias{so_ToolSettings}
\title{so_ToolSettings reference class}
\description{
	Reference Class for the SO/SOBlock/ToolSettings element of a PharmML-SO data structure
}
\section{Methods}{
so_ToolSettings$new() - Create a new empty so_ToolSettings object\cr
so_ToolSettings$add_File(object) - Add a File\cr
so_ToolSettings$remove_File(object, i) - Remove the File having index i\cr
}
\section{Fields}{
$File - A list of \link{so_ExternalFile} objects\cr
}
\keyword{so_ToolSettings}
