\name{so_VariabilityModel}
\alias{so_VariabilityModel}
\title{so_VariabilityModel reference class}
\description{
	Reference Class for the PharmML/ModelDefinition/VariabilityModel element of a PharmML-SO data structure
}
\section{Methods}{
so_VariabilityModel$new() - Create a new empty so_VariabilityModel object
}
\section{Fields}{
$Description - A character string\cr
$Name - A character string\cr
$Level - A \link{so_Level} object\cr
$id - A character string attribute\cr
$blkId - A character string attribute\cr
$type - A character string attribute\cr
}
\keyword{so_VariabilityModel}
